# src/main/openapi/kicker-rest.yml
swagger: "2.0"
info:
  description: "Kicker League API"
  version: "1.0.0"
  title: "Kicker League API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "kicker.internal.52north.org"
basePath: "/v1"
...
schemes:
- "https"
paths:
  /matches:
    get:
      tags:
      - "match"
      responses:
        200:
          description: "Matches, pageable"
          schema:
            $ref: "#/definitions/MatchesResponse"
    post:
      tags:
      - "match"
      parameters:
      - in: "body"
        name: "body"
        description: "Created match object"
        required: true
        schema:
          $ref: "#/definitions/Match"
      responses:
        200:
          description: "Create new matches"
          schema:
            $ref: "#/definitions/MatchesResponse"
  /matches/{id}:
    get:
      tags:
      - "match"
      parameters: 
      - name: "id"
        in: "path"
        description: "the match id"
        required: true
        type: "string"
      responses:
        200:
          description: "one Match"
          schema:
            $ref: "#/definitions/Match"
...
securityDefinitions:
  kicker_auth:
    type: "oauth2"
    authorizationUrl: "https://kicker.int.52north.org/oauth/dialog"
    flow: "implicit"
    scopes:
      write:player: "modify your account"
      read:player: "read your account"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Paging:
    type: "object"
    properties:
      total:
        type: "integer"
        format: "int32"
      page:
        type: "integer"
        format: "int32"
      size:
        type: "integer"
        format: "int32"
    required: 
      - $all
  MatchesResponse:
    type: "object"
    properties:
      paging:
        $ref: "#/definitions/Paging"
      matches:
        type: "array"
        items:
          $ref: "#/definitions/Match"
    required: 
      - $all
  Match:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      home:
        $ref: "#/definitions/Player"
      guest:
        $ref: "#/definitions/Player"
      score:
        $ref: "#/definitions/Score"
      events:
        type: "array"
        items:
          $ref: "#/definitions/MatchEvent"
    required: 
      - home
      - guest
  Score:
    type: "object"
    properties:
      home:
        type: integer
      guest:
        type: integer
    required: 
      - $all
  MatchEvent:
    type: "object"
    properties:
      dateTime:
        type: "object"
        format: "date-time"
      score:
        $ref: "#/definitions/Score"
      halftime:
        type: "boolean"
      fulltime:
        type: "boolean"
      fulltime:
        type: "boolean"
      finished:
        type: "boolean"
      pause:
        type: "boolean"
      resume:
        type: "boolean"
      cancel:
        type: "boolean"
    required: 
      - dateTime
  Player:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      nickName:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    required: 
      - id
      - username